Fix stuffs

+++ src/main/scala/decompositions/CostModel.scala
@@ -3,11 +3,11 @@ package decompositions
 import sql.Relation
 
 object CostModel {
-	def getJoinOpCost(lhs: PlanNode, rhs: PlanNode)(implicit cardinalities: Map[Set[Relation], Long] = Map.empty): Double = {
+	def getJoinOpCost(lhs: PlanNode, rhs: PlanNode)(implicit cardinalities: Map[Set[Relation], Double] = Map.empty): Double = {
 		lhs.cardinality + rhs.cardinality + cardinalities(lhs.allJoinedRelations ++ rhs.allJoinedRelations)
 	}
 
-	def getCumulativeCost(lhs: PlanNode, rhs: PlanNode): Long = {
+	def getCumulativeCost(lhs: PlanNode, rhs: PlanNode): Double = {
 		lhs.cumulativeCost + rhs.cumulativeCost + lhs.cardinality + rhs.cardinality
 	}
 }
